requirements.txt
streamlit>=1.24.0
pandas
numpy
plotly
matplotlib
psycopg2-binary
sqlalchemy
bcrypt==4.0.1
